requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
APScheduler==3.10.4
Pillow==10.1.0
